server/pyproject.toml
[project]
name = "mamahealth"
version = "0.1.0"
description = "Monorepo for the mama health coding challenge."
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[dependency-groups]
prod = [
    "fastapi[standard]>=0.115,<1.0",
    "pydantic>=2.11,<3.0",
    "pydantic-settings>=2.9,<3.0",
    "sqlalchemy>=2.0,<3.0",
    "python-dotenv>=1.1,<2.0",
    "websockets>=15.0,<16.0",
]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-asyncio>=0.26,<1.0",
    "pytest-cov>=6.1,<7.0",
    "ruff>=0.11,<1.0",
    "mypy>=1.15,<2.0",
    "pre-commit>=4.2,<5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
default-groups = ["dev"]
package = false

[tool.fastapi]
entrypoint = "main:app"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["."]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
plugins = ["pydantic.mypy"]
